[feature/2023-01-21-1] refactor security configs
