+++ internal/admin/delete_group.go
@@ -1,5 +1,10 @@
 package admin
 
 func (a *Admin) DeleteGroup(group string) error {
-	return a.getSaramaAdmin().DeleteConsumerGroup(group)
+	adm, err := a.getSaramaAdmin()
+	if err != nil {
+		return err
+	}
+
+	return adm.DeleteConsumerGroup(group)
 }
