+++ internal/sarama_config/token_provider.go
@@ -1,4 +1,4 @@
-package utils
+package sarama_config
 
 import (
 	"context"
